+++ CHANGELOG.md
@@ -12,6 +12,7 @@ baseline.
 ### Changed
 
 - The "What's New" dialog now features a scroll bar for longer entries
+- If the "What's New" dialog contains more than 5 entries, a More button accompanied by a downward-pointing arrow icon will appear beneath the final entry.
 
 ## [1.6.2]